объеденены модальные окна завершения игры
