+++ packages/client/src/components/index.ts
@@ -1,2 +1,3 @@
 export { default as Header } from './Header/Header'
 export { default as Modal } from './Modal/Modal'
+export { default as GameDialog } from './GameDialog/GameDialog'
